pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlproj"
version = "0.1.0"
dependencies = [
  "numpy",
  "pandas",
  "scikit-learn>=1.2",
  "matplotlib",
  "pyyaml",
  "joblib",
  "lightgbm>=4.0; platform_system!='Darwin' or platform_machine!='arm64'",
  "catboost>=1.2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]  # чтобы пакеты внутри src подтянулись